MBTA-51: Refactor database.py to enable extensibility
We need to make it easy for different parts of the
codebase to hook into the datbase models.  Now, all
you have to do to use the database is import
init_db, and whichever models you want.  run init_db,
and use the models to query.  If you want to insert/delete
you just import the DB_SESSION and use that.