MBTA-61: Hookup Inputs to AppState & send to API
I've hooked up all the inputs to the application state, which
means that if you click on any of the checkboxes, or set the
date/time ranges, or click any of the buttons, then it will
set the appropriate fields in the the application state.

I've also made it so that the page doesn't load data immediately,
but also setup the date/time ranges such that they're somewhat
useful.  The date/time params get sent to the alert_events API
in the GET request.

When the table doesn't have any data, it displays a
"No Data Loaded" message.  When the date/time ranges haven't
been set, you can't request data from the API, because the button
is disabled.

The next step will be creating a component that recognizes changes
to the filters and filters the data appropriately.  After that
we can just render the filtered data in the table.