Disable Automatic Webapp Deployment
The instance is way too small to run to run the webpack
optimize script.  When I run it on my local computer, it
takes about 30 seconds, whereas when it runs via Jenkins
it takes over an hour and usually kills Jenkins via OOM.

This isn't an issue with the script, but we should really
have a larger build slave.  Moving to manual deployment
of the webapp or the moment, since it's not worth the
effort.