MBTA-61: Default Filter Checks to ON
The general assumption is that data should not be filtered
out if the corresponding filter box is checked.  eg. If
I'm filtering on line and the inbound box is checked but
the outbound box isn't then the table will only contain
data that is inbound.

For this to work, we need to default all the checkboxes
to on, since people also want to see all the data at first.