MBTA-61: First attempt at Dashboard filtering
This PR integrates the filter state controlled by the toggle
buttons in the Dashboard with the data.  The end result is that
the data is filtered for other parts of the application to use,
and the table renders filtered data.

It seems like it works, but I had to make a couple of assumptions
about the formatting of dates/times/lines/etc.

I also decided that the logic for the alert filters should be
OR for related items (eg. Timely vs. Late) but AND between
unrelated items (eg. Timely vs. Existed).  This means that
it won't be possible to query for alerts that were Nonexistent
OR Late OR Innaccurate, which might be needed.  We need to
give up a degree of flexibility unless we plan on writing
a logic parsing system.

I'm also considering the best way to handle unexpected route names.
My current method is to disable filtering my route if the route
isn't expected, but it might be worth adding unexpected route
names to the filtering toggles, so they can be filtered.  Either
way, I don't want malformed data to disappear.  It needs to be
as visible as possible because it either needs to be fixed, or
it's legitimate data.

I plan to follow this up, once we can test it on real data.