Catch some errors that weren't being caught
It seems like the archiver hadn't been running for a while, because
it encountered a Integrity error with a duplicate key.  I had hoped
we were done with duplicate key issues, but it seems like there's
still a way to go.  At least this way, it won't kill the entire
application.